refactor(whatsrust): extract event models
